Add various work arounds for testing of INTERNAL wrapper
The testing framework from simphony-common uses default values
for each value in the DataContainer. For some CUBA (i.e.
material), we need a specific value that corresponds to how
the engine is configured (i.e. what materials it contains). So
some tests are overriden to ensure that the test particles are
configured correctly.